error fixes, code style
