+++ src/components/dataclasses.py
@@ -9,7 +9,7 @@ class Layout:
     def __init__(
         self,
         x_coordinate: float,
-        y_coordinate,
+        y_coordinate: float,
     ):
         self.x_coordinate = x_coordinate
         self.y_coordinate = y_coordinate
