+++ src/metrics/job_metrics.py
@@ -1,4 +1,4 @@
-from src.job_execution.job_execution_handler import JobStatus
+from src.job_execution.job_status import JobStatus
 from src.metrics.base_metrics import Metrics
 from datetime import datetime, timedelta
